started work on code clean up and module support
