+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/thrasher-corp/go-talib
+
+go 1.13
+
+require github.com/stretchr/testify v1.5.1
