+++ indicators/indicators.go
@@ -1,0 +1,504 @@
+package indicators
+
+import "math"
+
+/* Math Transform Functions */
+
+// Acos - Vector Trigonometric ACOS
+func Acos(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Acos(inReal[i])
+	}
+	return outReal
+}
+
+// Asin - Vector Trigonometric ASIN
+func Asin(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Asin(inReal[i])
+	}
+	return outReal
+}
+
+// Atan - Vector Trigonometric ATAN
+func Atan(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Atan(inReal[i])
+	}
+	return outReal
+}
+
+// Ceil - Vector CEIL
+func Ceil(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Ceil(inReal[i])
+	}
+	return outReal
+}
+
+// Cos - Vector Trigonometric COS
+func Cos(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Cos(inReal[i])
+	}
+	return outReal
+}
+
+// Cosh - Vector Trigonometric COSH
+func Cosh(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Cosh(inReal[i])
+	}
+	return outReal
+}
+
+// Exp - Vector atrithmetic EXP
+func Exp(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Exp(inReal[i])
+	}
+	return outReal
+}
+
+// Floor - Vector FLOOR
+func Floor(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Floor(inReal[i])
+	}
+	return outReal
+}
+
+// Ln - Vector natural log LN
+func Ln(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Log(inReal[i])
+	}
+	return outReal
+}
+
+// Log10 - Vector LOG10
+func Log10(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Log10(inReal[i])
+	}
+	return outReal
+}
+
+// Sin - Vector Trigonometric SIN
+func Sin(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Sin(inReal[i])
+	}
+	return outReal
+}
+
+// Sinh - Vector Trigonometric SINH
+func Sinh(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Sinh(inReal[i])
+	}
+	return outReal
+}
+
+// Sqrt - Vector SQRT
+func Sqrt(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Sqrt(inReal[i])
+	}
+	return outReal
+}
+
+// Tan - Vector Trigonometric TAN
+func Tan(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Tan(inReal[i])
+	}
+	return outReal
+}
+
+// Tanh - Vector Trigonometric TANH
+func Tanh(inReal []float64) []float64 {
+	outReal := make([]float64, len(inReal))
+	for i := 0; i < len(inReal); i++ {
+		outReal[i] = math.Tanh(inReal[i])
+	}
+	return outReal
+}
+
+/* Math Operator Functions */
+
+// Add - Vector arithmetic addition
+func Add(inReal0 []float64, inReal1 []float64) []float64 {
+	outReal := make([]float64, len(inReal0))
+	for i := 0; i < len(inReal0); i++ {
+		outReal[i] = inReal0[i] + inReal1[i]
+	}
+	return outReal
+}
+
+// Div - Vector arithmetic division
+func Div(inReal0 []float64, inReal1 []float64) []float64 {
+	outReal := make([]float64, len(inReal0))
+	for i := 0; i < len(inReal0); i++ {
+		outReal[i] = inReal0[i] / inReal1[i]
+	}
+	return outReal
+}
+
+// Max - Highest value over a period
+func Max(inReal []float64, inTimePeriod int) []float64 {
+
+	outReal := make([]float64, len(inReal))
+
+	if inTimePeriod < 2 {
+		return outReal
+	}
+
+	nbInitialElementNeeded := inTimePeriod - 1
+	startIdx := nbInitialElementNeeded
+	outIdx := startIdx
+	today := startIdx
+	trailingIdx := startIdx - nbInitialElementNeeded
+	highestIdx := -1
+	highest := 0.0
+
+	for today < len(outReal) {
+
+		tmp := inReal[today]
+
+		if highestIdx < trailingIdx {
+			highestIdx = trailingIdx
+			highest = inReal[highestIdx]
+			i := highestIdx + 1
+			for i <= today {
+				tmp = inReal[i]
+				if tmp > highest {
+					highestIdx = i
+					highest = tmp
+				}
+				i++
+			}
+		} else if tmp >= highest {
+			highestIdx = today
+			highest = tmp
+		}
+		outReal[outIdx] = highest
+		outIdx++
+		trailingIdx++
+		today++
+	}
+
+	return outReal
+}
+
+// MaxIndex - Index of highest value over a specified period
+func MaxIndex(inReal []float64, inTimePeriod int) []float64 {
+
+	outReal := make([]float64, len(inReal))
+
+	if inTimePeriod < 2 {
+		return outReal
+	}
+
+	nbInitialElementNeeded := inTimePeriod - 1
+	startIdx := nbInitialElementNeeded
+	outIdx := startIdx
+	today := startIdx
+	trailingIdx := startIdx - nbInitialElementNeeded
+	highestIdx := -1
+	highest := 0.0
+	for today < len(inReal) {
+		tmp := inReal[today]
+		if highestIdx < trailingIdx {
+			highestIdx = trailingIdx
+			highest = inReal[highestIdx]
+			i := highestIdx + 1
+			for i <= today {
+				tmp := inReal[i]
+				if tmp > highest {
+					highestIdx = i
+					highest = tmp
+				}
+				i++
+			}
+		} else if tmp >= highest {
+			highestIdx = today
+			highest = tmp
+		}
+		outReal[outIdx] = float64(highestIdx)
+		outIdx++
+		trailingIdx++
+		today++
+	}
+
+	return outReal
+}
+
+// Min - Lowest value over a period
+func Min(inReal []float64, inTimePeriod int) []float64 {
+
+	outReal := make([]float64, len(inReal))
+
+	if inTimePeriod < 2 {
+		return outReal
+	}
+
+	nbInitialElementNeeded := inTimePeriod - 1
+	startIdx := nbInitialElementNeeded
+	outIdx := startIdx
+	today := startIdx
+	trailingIdx := startIdx - nbInitialElementNeeded
+	lowestIdx := -1
+	lowest := 0.0
+	for today < len(outReal) {
+
+		tmp := inReal[today]
+
+		if lowestIdx < trailingIdx {
+			lowestIdx = trailingIdx
+			lowest = inReal[lowestIdx]
+			i := lowestIdx + 1
+			for i <= today {
+				tmp = inReal[i]
+				if tmp < lowest {
+					lowestIdx = i
+					lowest = tmp
+				}
+				i++
+			}
+		} else if tmp <= lowest {
+			lowestIdx = today
+			lowest = tmp
+		}
+		outReal[outIdx] = lowest
+		outIdx++
+		trailingIdx++
+		today++
+	}
+
+	return outReal
+}
+
+// MinIndex - Index of lowest value over a specified period
+func MinIndex(inReal []float64, inTimePeriod int) []float64 {
+
+	outReal := make([]float64, len(inReal))
+
+	if inTimePeriod < 2 {
+		return outReal
+	}
+
+	nbInitialElementNeeded := inTimePeriod - 1
+	startIdx := nbInitialElementNeeded
+	outIdx := startIdx
+	today := startIdx
+	trailingIdx := startIdx - nbInitialElementNeeded
+	lowestIdx := -1
+	lowest := 0.0
+	for today < len(inReal) {
+		tmp := inReal[today]
+		if lowestIdx < trailingIdx {
+			lowestIdx = trailingIdx
+			lowest = inReal[lowestIdx]
+			i := lowestIdx + 1
+			for i <= today {
+				tmp = inReal[i]
+				if tmp < lowest {
+					lowestIdx = i
+					lowest = tmp
+				}
+				i++
+			}
+		} else if tmp <= lowest {
+			lowestIdx = today
+			lowest = tmp
+		}
+		outReal[outIdx] = float64(lowestIdx)
+		outIdx++
+		trailingIdx++
+		today++
+	}
+	return outReal
+}
+
+// MinMax - Lowest and highest values over a specified period
+func MinMax(inReal []float64, inTimePeriod int) ([]float64, []float64) {
+
+	outMin := make([]float64, len(inReal))
+	outMax := make([]float64, len(inReal))
+
+	nbInitialElementNeeded := (inTimePeriod - 1)
+	startIdx := nbInitialElementNeeded
+	outIdx := startIdx
+	today := startIdx
+	trailingIdx := startIdx - nbInitialElementNeeded
+	highestIdx := -1
+	highest := 0.0
+	lowestIdx := -1
+	lowest := 0.0
+	for today < len(inReal) {
+		tmpLow, tmpHigh := inReal[today], inReal[today]
+		if highestIdx < trailingIdx {
+			highestIdx = trailingIdx
+			highest = inReal[highestIdx]
+			i := highestIdx
+			i++
+			for i <= today {
+				tmpHigh = inReal[i]
+				if tmpHigh > highest {
+					highestIdx = i
+					highest = tmpHigh
+				}
+				i++
+			}
+		} else if tmpHigh >= highest {
+			highestIdx = today
+			highest = tmpHigh
+		}
+		if lowestIdx < trailingIdx {
+			lowestIdx = trailingIdx
+			lowest = inReal[lowestIdx]
+			i := lowestIdx
+			i++
+			for i <= today {
+				tmpLow = inReal[i]
+				if tmpLow < lowest {
+					lowestIdx = i
+					lowest = tmpLow
+				}
+				i++
+			}
+		} else if tmpLow <= lowest {
+			lowestIdx = today
+			lowest = tmpLow
+		}
+		outMax[outIdx] = highest
+		outMin[outIdx] = lowest
+		outIdx++
+		trailingIdx++
+		today++
+	}
+	return outMin, outMax
+}
+
+// MinMaxIndex - Indexes of lowest and highest values over a specified period
+func MinMaxIndex(inReal []float64, inTimePeriod int) ([]float64, []float64) {
+
+	outMinIdx := make([]float64, len(inReal))
+	outMaxIdx := make([]float64, len(inReal))
+
+	nbInitialElementNeeded := (inTimePeriod - 1)
+	startIdx := nbInitialElementNeeded
+	outIdx := startIdx
+	today := startIdx
+	trailingIdx := startIdx - nbInitialElementNeeded
+	highestIdx := -1
+	highest := 0.0
+	lowestIdx := -1
+	lowest := 0.0
+	for today < len(inReal) {
+		tmpLow, tmpHigh := inReal[today], inReal[today]
+		if highestIdx < trailingIdx {
+			highestIdx = trailingIdx
+			highest = inReal[highestIdx]
+			i := highestIdx
+			i++
+			for i <= today {
+				tmpHigh = inReal[i]
+				if tmpHigh > highest {
+					highestIdx = i
+					highest = tmpHigh
+				}
+				i++
+			}
+		} else if tmpHigh >= highest {
+			highestIdx = today
+			highest = tmpHigh
+		}
+		if lowestIdx < trailingIdx {
+			lowestIdx = trailingIdx
+			lowest = inReal[lowestIdx]
+			i := lowestIdx
+			i++
+			for i <= today {
+				tmpLow = inReal[i]
+				if tmpLow < lowest {
+					lowestIdx = i
+					lowest = tmpLow
+				}
+				i++
+			}
+		} else if tmpLow <= lowest {
+			lowestIdx = today
+			lowest = tmpLow
+		}
+		outMaxIdx[outIdx] = float64(highestIdx)
+		outMinIdx[outIdx] = float64(lowestIdx)
+		outIdx++
+		trailingIdx++
+		today++
+	}
+	return outMinIdx, outMaxIdx
+}
+
+// Mult - Vector arithmetic multiply
+func Mult(inReal0 []float64, inReal1 []float64) []float64 {
+	outReal := make([]float64, len(inReal0))
+	for i := 0; i < len(inReal0); i++ {
+		outReal[i] = inReal0[i] * inReal1[i]
+	}
+	return outReal
+}
+
+// Sub - Vector arithmetic subtraction
+func Sub(inReal0 []float64, inReal1 []float64) []float64 {
+	outReal := make([]float64, len(inReal0))
+	for i := 0; i < len(inReal0); i++ {
+		outReal[i] = inReal0[i] - inReal1[i]
+	}
+	return outReal
+}
+
+// Sum - Vector summation
+func Sum(inReal []float64, inTimePeriod int) []float64 {
+
+	outReal := make([]float64, len(inReal))
+
+	lookbackTotal := inTimePeriod - 1
+	startIdx := lookbackTotal
+	periodTotal := 0.0
+	trailingIdx := startIdx - lookbackTotal
+	i := trailingIdx
+	if inTimePeriod > 1 {
+		for i < startIdx {
+			periodTotal += inReal[i]
+			i++
+		}
+	}
+	outIdx := startIdx
+	for i < len(inReal) {
+		periodTotal += inReal[i]
+		tempReal := periodTotal
+		periodTotal -= inReal[trailingIdx]
+		outReal[outIdx] = tempReal
+		i++
+		trailingIdx++
+		outIdx++
+	}
+
+	return outReal
+}
