+++ indicators/talib.go
@@ -1,1444 +1,17 @@
 /*
 Copyright 2016 Mark Chenoweth
+Copyright 2020 The GoCryptoTrader Developers
 Licensed under terms of MIT license (see LICENSE)
 */
 
 // Package talib is a pure Go port of TA-Lib (http://ta-lib.org) Technical Analysis Library
-package talib
+package indicators
 
 import (
 	"errors"
 	"math"
 )
 
-// MaType - Moving average type
-type MaType int
-
-type moneyFlow struct {
-	positive float64
-	negative float64
-}
-
-// Kinds of moving averages
-const (
-	SMA MaType = iota
-	EMA
-	WMA
-	DEMA
-	TEMA
-	TRIMA
-	KAMA
-	MAMA
-	T3MA
-)
-
-/* Overlap Studies */
-
-// BBands - Bollinger Bands
-// upperband, middleband, lowerband = BBands(close, timeperiod=5, nbdevup=2, nbdevdn=2, matype=0)
-func BBands(inReal []float64, inTimePeriod int, inNbDevUp float64, inNbDevDn float64, inMAType MaType) ([]float64, []float64, []float64) {
-
-	outRealUpperBand := make([]float64, len(inReal))
-	outRealMiddleBand := Ma(inReal, inTimePeriod, inMAType)
-	outRealLowerBand := make([]float64, len(inReal))
-
-	tempBuffer2 := StdDev(inReal, inTimePeriod, 1.0)
-
-	if inNbDevUp == inNbDevDn {
-
-		if inNbDevUp == 1.0 {
-			for i := 0; i < len(inReal); i++ {
-				tempReal := tempBuffer2[i]
-				tempReal2 := outRealMiddleBand[i]
-				outRealUpperBand[i] = tempReal2 + tempReal
-				outRealLowerBand[i] = tempReal2 - tempReal
-			}
-		} else {
-			for i := 0; i < len(inReal); i++ {
-				tempReal := tempBuffer2[i] * inNbDevUp
-				tempReal2 := outRealMiddleBand[i]
-				outRealUpperBand[i] = tempReal2 + tempReal
-				outRealLowerBand[i] = tempReal2 - tempReal
-			}
-		}
-	} else if inNbDevUp == 1.0 {
-		for i := 0; i < len(inReal); i++ {
-			tempReal := tempBuffer2[i]
-			tempReal2 := outRealMiddleBand[i]
-			outRealUpperBand[i] = tempReal2 + tempReal
-			outRealLowerBand[i] = tempReal2 - (tempReal * inNbDevDn)
-		}
-	} else if inNbDevDn == 1.0 {
-		for i := 0; i < len(inReal); i++ {
-			tempReal := tempBuffer2[i]
-			tempReal2 := outRealMiddleBand[i]
-			outRealLowerBand[i] = tempReal2 - tempReal
-			outRealUpperBand[i] = tempReal2 + (tempReal * inNbDevUp)
-		}
-	} else {
-		for i := 0; i < len(inReal); i++ {
-			tempReal := tempBuffer2[i]
-			tempReal2 := outRealMiddleBand[i]
-			outRealUpperBand[i] = tempReal2 + (tempReal * inNbDevUp)
-			outRealLowerBand[i] = tempReal2 - (tempReal * inNbDevDn)
-		}
-	}
-	return outRealUpperBand, outRealMiddleBand, outRealLowerBand
-}
-
-// Dema - Double Exponential Moving Average
-func Dema(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-	firstEMA := Ema(inReal, inTimePeriod)
-	secondEMA := Ema(firstEMA[inTimePeriod-1:], inTimePeriod)
-
-	for outIdx, secondEMAIdx := (inTimePeriod*2)-2, inTimePeriod-1; outIdx < len(inReal); outIdx, secondEMAIdx = outIdx+1, secondEMAIdx+1 {
-		outReal[outIdx] = (2.0 * firstEMA[outIdx]) - secondEMA[secondEMAIdx]
-	}
-
-	return outReal
-}
-
-// Ema - Exponential Moving Average
-func ema(inReal []float64, inTimePeriod int, k1 float64) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	lookbackTotal := inTimePeriod - 1
-	startIdx := lookbackTotal
-	today := startIdx - lookbackTotal
-	i := inTimePeriod
-	tempReal := 0.0
-	for i > 0 {
-		tempReal += inReal[today]
-		today++
-		i--
-	}
-
-	prevMA := tempReal / float64(inTimePeriod)
-
-	for today <= startIdx {
-		prevMA = ((inReal[today] - prevMA) * k1) + prevMA
-		today++
-	}
-	outReal[startIdx] = prevMA
-	outIdx := startIdx + 1
-	for today < len(inReal) {
-		prevMA = ((inReal[today] - prevMA) * k1) + prevMA
-		outReal[outIdx] = prevMA
-		today++
-		outIdx++
-	}
-
-	return outReal
-}
-
-// Ema - Exponential Moving Average
-func Ema(inReal []float64, inTimePeriod int) []float64 {
-
-	k := 2.0 / float64(inTimePeriod+1)
-	outReal := ema(inReal, inTimePeriod, k)
-	return outReal
-}
-
-// HtTrendline - Hilbert Transform - Instantaneous Trendline (lookback=63)
-func HtTrendline(inReal []float64) []float64 {
-
-	outReal := make([]float64, len(inReal))
-	a := 0.0962
-	b := 0.5769
-	detrenderOdd := make([]float64, 3)
-	detrenderEven := make([]float64, 3)
-	q1Odd := make([]float64, 3)
-	q1Even := make([]float64, 3)
-	jIOdd := make([]float64, 3)
-	jIEven := make([]float64, 3)
-	jQOdd := make([]float64, 3)
-	jQEven := make([]float64, 3)
-	smoothPriceIdx := 0
-	maxIdxSmoothPrice := (50 - 1)
-	smoothPrice := make([]float64, maxIdxSmoothPrice+1)
-	iTrend1 := 0.0
-	iTrend2 := 0.0
-	iTrend3 := 0.0
-	tempReal := math.Atan(1)
-	rad2Deg := 45.0 / tempReal
-	lookbackTotal := 63
-	startIdx := lookbackTotal
-	trailingWMAIdx := startIdx - lookbackTotal
-	today := trailingWMAIdx
-	tempReal = inReal[today]
-	today++
-	periodWMASub := tempReal
-	periodWMASum := tempReal
-	tempReal = inReal[today]
-	today++
-	periodWMASub += tempReal
-	periodWMASum += tempReal * 2.0
-	tempReal = inReal[today]
-	today++
-	periodWMASub += tempReal
-	periodWMASum += tempReal * 3.0
-	trailingWMAValue := 0.0
-	i := 34
-	for ok := true; ok; {
-		tempReal = inReal[today]
-		today++
-		periodWMASub += tempReal
-		periodWMASub -= trailingWMAValue
-		periodWMASum += tempReal * 4.0
-		trailingWMAValue = inReal[trailingWMAIdx]
-		trailingWMAIdx++
-		//smoothedValue := periodWMASum * 0.1
-		periodWMASum -= periodWMASub
-		i--
-		ok = i != 0
-	}
-	hilbertIdx := 0
-	detrender := 0.0
-	prevDetrenderOdd := 0.0
-	prevDetrenderEven := 0.0
-	prevDetrenderInputOdd := 0.0
-	prevDetrenderInputEven := 0.0
-	q1 := 0.0
-	prevq1Odd := 0.0
-	prevq1Even := 0.0
-	prevq1InputOdd := 0.0
-	prevq1InputEven := 0.0
-	jI := 0.0
-	prevJIOdd := 0.0
-	prevJIEven := 0.0
-	prevJIInputOdd := 0.0
-	prevJIInputEven := 0.0
-	jQ := 0.0
-	prevJQOdd := 0.0
-	prevJQEven := 0.0
-	prevJQInputOdd := 0.0
-	prevJQInputEven := 0.0
-	period := 0.0
-	outIdx := 63
-	previ2 := 0.0
-	prevq2 := 0.0
-	Re := 0.0
-	Im := 0.0
-	i1ForOddPrev3 := 0.0
-	i1ForEvenPrev3 := 0.0
-	i1ForOddPrev2 := 0.0
-	i1ForEvenPrev2 := 0.0
-	smoothPeriod := 0.0
-	q2 := 0.0
-	i2 := 0.0
-	for today < len(inReal) {
-		adjustedPrevPeriod := (0.075 * period) + 0.54
-		todayValue := inReal[today]
-		periodWMASub += todayValue
-		periodWMASub -= trailingWMAValue
-		periodWMASum += todayValue * 4.0
-		trailingWMAValue = inReal[trailingWMAIdx]
-		trailingWMAIdx++
-		smoothedValue := periodWMASum * 0.1
-		periodWMASum -= periodWMASub
-		smoothPrice[smoothPriceIdx] = smoothedValue
-		if (today % 2) == 0 {
-			hilbertTempReal := a * smoothedValue
-			detrender = -detrenderEven[hilbertIdx]
-			detrenderEven[hilbertIdx] = hilbertTempReal
-			detrender += hilbertTempReal
-			detrender -= prevDetrenderEven
-			prevDetrenderEven = b * prevDetrenderInputEven
-			detrender += prevDetrenderEven
-			prevDetrenderInputEven = smoothedValue
-			detrender *= adjustedPrevPeriod
-			hilbertTempReal = a * detrender
-			q1 = -q1Even[hilbertIdx]
-			q1Even[hilbertIdx] = hilbertTempReal
-			q1 += hilbertTempReal
-			q1 -= prevq1Even
-			prevq1Even = b * prevq1InputEven
-			q1 += prevq1Even
-			prevq1InputEven = detrender
-			q1 *= adjustedPrevPeriod
-			hilbertTempReal = a * i1ForEvenPrev3
-			jI = -jIEven[hilbertIdx]
-			jIEven[hilbertIdx] = hilbertTempReal
-			jI += hilbertTempReal
-			jI -= prevJIEven
-			prevJIEven = b * prevJIInputEven
-			jI += prevJIEven
-			prevJIInputEven = i1ForEvenPrev3
-			jI *= adjustedPrevPeriod
-			hilbertTempReal = a * q1
-			jQ = -jQEven[hilbertIdx]
-			jQEven[hilbertIdx] = hilbertTempReal
-			jQ += hilbertTempReal
-			jQ -= prevJQEven
-			prevJQEven = b * prevJQInputEven
-			jQ += prevJQEven
-			prevJQInputEven = q1
-			jQ *= adjustedPrevPeriod
-			hilbertIdx++
-			if hilbertIdx == 3 {
-				hilbertIdx = 0
-			}
-			q2 = (0.2 * (q1 + jI)) + (0.8 * prevq2)
-			i2 = (0.2 * (i1ForEvenPrev3 - jQ)) + (0.8 * previ2)
-			i1ForOddPrev3 = i1ForOddPrev2
-			i1ForOddPrev2 = detrender
-		} else {
-			hilbertTempReal := a * smoothedValue
-			detrender = -detrenderOdd[hilbertIdx]
-			detrenderOdd[hilbertIdx] = hilbertTempReal
-			detrender += hilbertTempReal
-			detrender -= prevDetrenderOdd
-			prevDetrenderOdd = b * prevDetrenderInputOdd
-			detrender += prevDetrenderOdd
-			prevDetrenderInputOdd = smoothedValue
-			detrender *= adjustedPrevPeriod
-			hilbertTempReal = a * detrender
-			q1 = -q1Odd[hilbertIdx]
-			q1Odd[hilbertIdx] = hilbertTempReal
-			q1 += hilbertTempReal
-			q1 -= prevq1Odd
-			prevq1Odd = b * prevq1InputOdd
-			q1 += prevq1Odd
-			prevq1InputOdd = detrender
-			q1 *= adjustedPrevPeriod
-			hilbertTempReal = a * i1ForOddPrev3
-			jI = -jIOdd[hilbertIdx]
-			jIOdd[hilbertIdx] = hilbertTempReal
-			jI += hilbertTempReal
-			jI -= prevJIOdd
-			prevJIOdd = b * prevJIInputOdd
-			jI += prevJIOdd
-			prevJIInputOdd = i1ForOddPrev3
-			jI *= adjustedPrevPeriod
-			hilbertTempReal = a * q1
-			jQ = -jQOdd[hilbertIdx]
-			jQOdd[hilbertIdx] = hilbertTempReal
-			jQ += hilbertTempReal
-			jQ -= prevJQOdd
-			prevJQOdd = b * prevJQInputOdd
-			jQ += prevJQOdd
-			prevJQInputOdd = q1
-			jQ *= adjustedPrevPeriod
-			q2 = (0.2 * (q1 + jI)) + (0.8 * prevq2)
-			i2 = (0.2 * (i1ForOddPrev3 - jQ)) + (0.8 * previ2)
-			i1ForEvenPrev3 = i1ForEvenPrev2
-			i1ForEvenPrev2 = detrender
-		}
-		Re = (0.2 * ((i2 * previ2) + (q2 * prevq2))) + (0.8 * Re)
-		Im = (0.2 * ((i2 * prevq2) - (q2 * previ2))) + (0.8 * Im)
-		prevq2 = q2
-		previ2 = i2
-		tempReal = period
-		if (Im != 0.0) && (Re != 0.0) {
-			period = 360.0 / (math.Atan(Im/Re) * rad2Deg)
-		}
-		tempReal2 := 1.5 * tempReal
-		if period > tempReal2 {
-			period = tempReal2
-		}
-		tempReal2 = 0.67 * tempReal
-		if period < tempReal2 {
-			period = tempReal2
-		}
-		if period < 6 {
-			period = 6
-		} else if period > 50 {
-			period = 50
-		}
-		period = (0.2 * period) + (0.8 * tempReal)
-		smoothPeriod = (0.33 * period) + (0.67 * smoothPeriod)
-		DCPeriod := smoothPeriod + 0.5
-		DCPeriodInt := math.Floor(DCPeriod)
-		idx := today
-		tempReal = 0.0
-		for i := 0; i < int(DCPeriodInt); i++ {
-			tempReal += inReal[idx]
-			idx--
-		}
-		if DCPeriodInt > 0 {
-			tempReal = tempReal / (DCPeriodInt * 1.0)
-		}
-		tempReal2 = (4.0*tempReal + 3.0*iTrend1 + 2.0*iTrend2 + iTrend3) / 10.0
-		iTrend3 = iTrend2
-		iTrend2 = iTrend1
-		iTrend1 = tempReal
-		if today >= startIdx {
-			outReal[outIdx] = tempReal2
-			outIdx++
-		}
-		smoothPriceIdx++
-		if smoothPriceIdx > maxIdxSmoothPrice {
-			smoothPriceIdx = 0
-		}
-
-		today++
-	}
-	return outReal
-}
-
-// Kama - Kaufman Adaptive Moving Average
-func Kama(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	constMax := 2.0 / (30.0 + 1.0)
-	constDiff := 2.0/(2.0+1.0) - constMax
-	lookbackTotal := inTimePeriod
-	startIdx := lookbackTotal
-	sumROC1 := 0.0
-	today := startIdx - lookbackTotal
-	trailingIdx := today
-	i := inTimePeriod
-	for i > 0 {
-		tempReal := inReal[today]
-		today++
-		tempReal -= inReal[today]
-		sumROC1 += math.Abs(tempReal)
-		i--
-	}
-	prevKAMA := inReal[today-1]
-	tempReal := inReal[today]
-	tempReal2 := inReal[trailingIdx]
-	trailingIdx++
-	periodROC := tempReal - tempReal2
-	trailingValue := tempReal2
-	if (sumROC1 <= periodROC) || (((-(0.00000000000001)) < sumROC1) && (sumROC1 < (0.00000000000001))) {
-		tempReal = 1.0
-	} else {
-		tempReal = math.Abs(periodROC / sumROC1)
-	}
-	tempReal = (tempReal * constDiff) + constMax
-	tempReal *= tempReal
-	prevKAMA = ((inReal[today] - prevKAMA) * tempReal) + prevKAMA
-	today++
-	for today <= startIdx {
-		tempReal = inReal[today]
-		tempReal2 = inReal[trailingIdx]
-		trailingIdx++
-		periodROC = tempReal - tempReal2
-		sumROC1 -= math.Abs(trailingValue - tempReal2)
-		sumROC1 += math.Abs(tempReal - inReal[today-1])
-		trailingValue = tempReal2
-		if (sumROC1 <= periodROC) || (((-(0.00000000000001)) < sumROC1) && (sumROC1 < (0.00000000000001))) {
-			tempReal = 1.0
-		} else {
-			tempReal = math.Abs(periodROC / sumROC1)
-		}
-		tempReal = (tempReal * constDiff) + constMax
-		tempReal *= tempReal
-		prevKAMA = ((inReal[today] - prevKAMA) * tempReal) + prevKAMA
-		today++
-	}
-	outReal[inTimePeriod] = prevKAMA
-	outIdx := inTimePeriod + 1
-	for today < len(inReal) {
-		tempReal = inReal[today]
-		tempReal2 = inReal[trailingIdx]
-		trailingIdx++
-		periodROC = tempReal - tempReal2
-		sumROC1 -= math.Abs(trailingValue - tempReal2)
-		sumROC1 += math.Abs(tempReal - inReal[today-1])
-		trailingValue = tempReal2
-		if (sumROC1 <= periodROC) || (((-(0.00000000000001)) < sumROC1) && (sumROC1 < (0.00000000000001))) {
-			tempReal = 1.0
-		} else {
-			tempReal = math.Abs(periodROC / sumROC1)
-		}
-		tempReal = (tempReal * constDiff) + constMax
-		tempReal *= tempReal
-		prevKAMA = ((inReal[today] - prevKAMA) * tempReal) + prevKAMA
-		today++
-		outReal[outIdx] = prevKAMA
-		outIdx++
-	}
-
-	return outReal
-}
-
-// Ma - Moving average
-func Ma(inReal []float64, inTimePeriod int, inMAType MaType) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	if inTimePeriod == 1 {
-		copy(outReal, inReal)
-		return outReal
-	}
-
-	switch inMAType {
-	case SMA:
-		outReal = Sma(inReal, inTimePeriod)
-	case EMA:
-		outReal = Ema(inReal, inTimePeriod)
-	case WMA:
-		outReal = Wma(inReal, inTimePeriod)
-	case DEMA:
-		outReal = Dema(inReal, inTimePeriod)
-	case TEMA:
-		outReal = Tema(inReal, inTimePeriod)
-	case TRIMA:
-		outReal = Trima(inReal, inTimePeriod)
-	case KAMA:
-		outReal = Kama(inReal, inTimePeriod)
-	case MAMA:
-		outReal, _ = Mama(inReal, 0.5, 0.05)
-	case T3MA:
-		outReal = T3(inReal, inTimePeriod, 0.7)
-	}
-	return outReal
-}
-
-// Mama - MESA Adaptive Moving Average (lookback=32)
-func Mama(inReal []float64, inFastLimit float64, inSlowLimit float64) ([]float64, []float64) {
-
-	outMAMA := make([]float64, len(inReal))
-	outFAMA := make([]float64, len(inReal))
-
-	a := 0.0962
-	b := 0.5769
-	detrenderOdd := make([]float64, 3)
-	detrenderEven := make([]float64, 3)
-	q1Odd := make([]float64, 3)
-	q1Even := make([]float64, 3)
-	jIOdd := make([]float64, 3)
-	jIEven := make([]float64, 3)
-	jQOdd := make([]float64, 3)
-	jQEven := make([]float64, 3)
-	rad2Deg := 180.0 / (4.0 * math.Atan(1))
-	lookbackTotal := 32
-	startIdx := lookbackTotal
-	trailingWMAIdx := startIdx - lookbackTotal
-	today := trailingWMAIdx
-	tempReal := inReal[today]
-	today++
-	periodWMASub := tempReal
-	periodWMASum := tempReal
-	tempReal = inReal[today]
-	today++
-	periodWMASub += tempReal
-	periodWMASum += tempReal * 2.0
-	tempReal = inReal[today]
-	today++
-	periodWMASub += tempReal
-	periodWMASum += tempReal * 3.0
-	trailingWMAValue := 0.0
-	i := 9
-	smoothedValue := 0.0
-	for ok := true; ok; {
-		tempReal = inReal[today]
-		today++
-		periodWMASub += tempReal
-		periodWMASub -= trailingWMAValue
-		periodWMASum += tempReal * 4.0
-		trailingWMAValue = inReal[trailingWMAIdx]
-		trailingWMAIdx++
-		smoothedValue = periodWMASum * 0.1
-		periodWMASum -= periodWMASub
-		i--
-		ok = i != 0
-	}
-	hilbertIdx := 0
-	detrenderOdd[0] = 0.0
-	detrenderOdd[1] = 0.0
-	detrenderOdd[2] = 0.0
-	detrenderEven[0] = 0.0
-	detrenderEven[1] = 0.0
-	detrenderEven[2] = 0.0
-	detrender := 0.0
-	prevDetrenderOdd := 0.0
-	prevDetrenderEven := 0.0
-	prevDetrenderInputOdd := 0.0
-	prevDetrenderInputEven := 0.0
-
-	q1Odd[0] = 0.0
-	q1Odd[1] = 0.0
-	q1Odd[2] = 0.0
-	q1Even[0] = 0.0
-	q1Even[1] = 0.0
-	q1Even[2] = 0.0
-	q1 := 0.0
-	prevq1Odd := 0.0
-	prevq1Even := 0.0
-	prevq1InputOdd := 0.0
-	prevq1InputEven := 0.0
-
-	jIOdd[0] = 0.0
-	jIOdd[1] = 0.0
-	jIOdd[2] = 0.0
-	jIEven[0] = 0.0
-	jIEven[1] = 0.0
-	jIEven[2] = 0.0
-	jI := 0.0
-	prevjIOdd := 0.0
-	prevjIEven := 0.0
-	prevjIInputOdd := 0.0
-	prevjIInputEven := 0.0
-
-	jQOdd[0] = 0.0
-	jQOdd[1] = 0.0
-	jQOdd[2] = 0.0
-	jQEven[0] = 0.0
-	jQEven[1] = 0.0
-	jQEven[2] = 0.0
-	jQ := 0.0
-	prevjQOdd := 0.0
-	prevjQEven := 0.0
-	prevjQInputOdd := 0.0
-	prevjQInputEven := 0.0
-
-	period := 0.0
-	outIdx := startIdx
-	previ2, prevq2 := 0.0, 0.0
-	Re, Im := 0.0, 0.0
-	mama, fama := 0.0, 0.0
-	i1ForOddPrev3, i1ForEvenPrev3 := 0.0, 0.0
-	i1ForOddPrev2, i1ForEvenPrev2 := 0.0, 0.0
-	prevPhase := 0.0
-	adjustedPrevPeriod := 0.0
-	todayValue := 0.0
-	hilbertTempReal := 0.0
-	for today < len(inReal) {
-		adjustedPrevPeriod = (0.075 * period) + 0.54
-		todayValue = inReal[today]
-
-		periodWMASub += todayValue
-		periodWMASub -= trailingWMAValue
-		periodWMASum += todayValue * 4.0
-		trailingWMAValue = inReal[trailingWMAIdx]
-		trailingWMAIdx++
-		smoothedValue = periodWMASum * 0.1
-		periodWMASum -= periodWMASub
-		q2, i2 := 0.0, 0.0
-		tempReal2 := 0.0
-		if (today % 2) == 0 {
-
-			hilbertTempReal = a * smoothedValue
-			detrender = -detrenderEven[hilbertIdx]
-			detrenderEven[hilbertIdx] = hilbertTempReal
-			detrender += hilbertTempReal
-			detrender -= prevDetrenderEven
-			prevDetrenderEven = b * prevDetrenderInputEven
-			detrender += prevDetrenderEven
-			prevDetrenderInputEven = smoothedValue
-			detrender *= adjustedPrevPeriod
-
-			hilbertTempReal = a * detrender
-			q1 = -q1Even[hilbertIdx]
-			q1Even[hilbertIdx] = hilbertTempReal
-			q1 += hilbertTempReal
-			q1 -= prevq1Even
-			prevq1Even = b * prevq1InputEven
-			q1 += prevq1Even
-			prevq1InputEven = detrender
-			q1 *= adjustedPrevPeriod
-
-			hilbertTempReal = a * i1ForEvenPrev3
-			jI = -jIEven[hilbertIdx]
-			jIEven[hilbertIdx] = hilbertTempReal
-			jI += hilbertTempReal
-			jI -= prevjIEven
-			prevjIEven = b * prevjIInputEven
-			jI += prevjIEven
-			prevjIInputEven = i1ForEvenPrev3
-			jI *= adjustedPrevPeriod
-
-			hilbertTempReal = a * q1
-			jQ = -jQEven[hilbertIdx]
-			jQEven[hilbertIdx] = hilbertTempReal
-			jQ += hilbertTempReal
-			jQ -= prevjQEven
-			prevjQEven = b * prevjQInputEven
-			jQ += prevjQEven
-			prevjQInputEven = q1
-			jQ *= adjustedPrevPeriod
-			hilbertIdx++
-			if hilbertIdx == 3 {
-				hilbertIdx = 0
-			}
-			q2 = (0.2 * (q1 + jI)) + (0.8 * prevq2)
-			i2 = (0.2 * (i1ForEvenPrev3 - jQ)) + (0.8 * previ2)
-			i1ForOddPrev3 = i1ForOddPrev2
-			i1ForOddPrev2 = detrender
-			if i1ForEvenPrev3 != 0.0 {
-				tempReal2 = (math.Atan(q1/i1ForEvenPrev3) * rad2Deg)
-			} else {
-				tempReal2 = 0.0
-			}
-		} else {
-
-			hilbertTempReal = a * smoothedValue
-			detrender = -detrenderOdd[hilbertIdx]
-			detrenderOdd[hilbertIdx] = hilbertTempReal
-			detrender += hilbertTempReal
-			detrender -= prevDetrenderOdd
-			prevDetrenderOdd = b * prevDetrenderInputOdd
-			detrender += prevDetrenderOdd
-			prevDetrenderInputOdd = smoothedValue
-			detrender *= adjustedPrevPeriod
-
-			hilbertTempReal = a * detrender
-			q1 = -q1Odd[hilbertIdx]
-			q1Odd[hilbertIdx] = hilbertTempReal
-			q1 += hilbertTempReal
-			q1 -= prevq1Odd
-			prevq1Odd = b * prevq1InputOdd
-			q1 += prevq1Odd
-			prevq1InputOdd = detrender
-			q1 *= adjustedPrevPeriod
-
-			hilbertTempReal = a * i1ForOddPrev3
-			jI = -jIOdd[hilbertIdx]
-			jIOdd[hilbertIdx] = hilbertTempReal
-			jI += hilbertTempReal
-			jI -= prevjIOdd
-			prevjIOdd = b * prevjIInputOdd
-			jI += prevjIOdd
-			prevjIInputOdd = i1ForOddPrev3
-			jI *= adjustedPrevPeriod
-
-			hilbertTempReal = a * q1
-			jQ = -jQOdd[hilbertIdx]
-			jQOdd[hilbertIdx] = hilbertTempReal
-			jQ += hilbertTempReal
-			jQ -= prevjQOdd
-			prevjQOdd = b * prevjQInputOdd
-			jQ += prevjQOdd
-			prevjQInputOdd = q1
-			jQ *= adjustedPrevPeriod
-
-			q2 = (0.2 * (q1 + jI)) + (0.8 * prevq2)
-			i2 = (0.2 * (i1ForOddPrev3 - jQ)) + (0.8 * previ2)
-			i1ForEvenPrev3 = i1ForEvenPrev2
-			i1ForEvenPrev2 = detrender
-			if i1ForOddPrev3 != 0.0 {
-				tempReal2 = (math.Atan(q1/i1ForOddPrev3) * rad2Deg)
-			} else {
-				tempReal2 = 0.0
-			}
-		}
-		tempReal = prevPhase - tempReal2
-		prevPhase = tempReal2
-		if tempReal < 1.0 {
-			tempReal = 1.0
-		}
-		if tempReal > 1.0 {
-			tempReal = inFastLimit / tempReal
-			if tempReal < inSlowLimit {
-				tempReal = inSlowLimit
-			}
-		} else {
-			tempReal = inFastLimit
-		}
-		mama = (tempReal * todayValue) + ((1 - tempReal) * mama)
-		tempReal *= 0.5
-		fama = (tempReal * mama) + ((1 - tempReal) * fama)
-		if today >= startIdx {
-			outMAMA[outIdx] = mama
-			outFAMA[outIdx] = fama
-			outIdx++
-		}
-		Re = (0.2 * ((i2 * previ2) + (q2 * prevq2))) + (0.8 * Re)
-		Im = (0.2 * ((i2 * prevq2) - (q2 * previ2))) + (0.8 * Im)
-		prevq2 = q2
-		previ2 = i2
-		tempReal = period
-		if (Im != 0.0) && (Re != 0.0) {
-			period = 360.0 / (math.Atan(Im/Re) * rad2Deg)
-		}
-		tempReal2 = 1.5 * tempReal
-		if period > tempReal2 {
-			period = tempReal2
-		}
-		tempReal2 = 0.67 * tempReal
-		if period < tempReal2 {
-			period = tempReal2
-		}
-		if period < 6 {
-			period = 6
-		} else if period > 50 {
-			period = 50
-		}
-		period = (0.2 * period) + (0.8 * tempReal)
-		today++
-	}
-	return outMAMA, outFAMA
-}
-
-// MaVp - Moving average with variable period
-func MaVp(inReal []float64, inPeriods []float64, inMinPeriod int, inMaxPeriod int, inMAType MaType) []float64 {
-
-	outReal := make([]float64, len(inReal))
-	startIdx := inMaxPeriod - 1
-	outputSize := len(inReal)
-
-	localPeriodArray := make([]float64, outputSize)
-	for i := startIdx; i < outputSize; i++ {
-		tempInt := int(inPeriods[i])
-		if tempInt < inMinPeriod {
-			tempInt = inMinPeriod
-		} else if tempInt > inMaxPeriod {
-			tempInt = inMaxPeriod
-		}
-		localPeriodArray[i] = float64(tempInt)
-	}
-
-	for i := startIdx; i < outputSize; i++ {
-		curPeriod := int(localPeriodArray[i])
-		if curPeriod != 0 {
-			localOutputArray := Ma(inReal, curPeriod, inMAType)
-			outReal[i] = localOutputArray[i]
-			for j := i + 1; j < outputSize; j++ {
-				if localPeriodArray[j] == float64(curPeriod) {
-					localPeriodArray[j] = 0
-					outReal[j] = localOutputArray[j]
-				}
-			}
-		}
-	}
-	return outReal
-}
-
-// MidPoint - MidPoint over period
-func MidPoint(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-	nbInitialElementNeeded := inTimePeriod - 1
-	startIdx := nbInitialElementNeeded
-	outIdx := inTimePeriod - 1
-	today := startIdx
-	trailingIdx := startIdx - nbInitialElementNeeded
-
-	for today < len(inReal) {
-		lowest := inReal[trailingIdx]
-		trailingIdx++
-		highest := lowest
-		for i := trailingIdx; i <= today; i++ {
-			tmp := inReal[i]
-			if tmp < lowest {
-				lowest = tmp
-			} else if tmp > highest {
-				highest = tmp
-			}
-		}
-		outReal[outIdx] = (highest + lowest) / 2.0
-		outIdx++
-		today++
-	}
-	return outReal
-}
-
-// MidPrice - Midpoint Price over period
-func MidPrice(inHigh []float64, inLow []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inHigh))
-
-	nbInitialElementNeeded := inTimePeriod - 1
-	startIdx := nbInitialElementNeeded
-	outIdx := inTimePeriod - 1
-	today := startIdx
-	trailingIdx := startIdx - nbInitialElementNeeded
-	for today < len(inHigh) {
-		lowest := inLow[trailingIdx]
-		highest := inHigh[trailingIdx]
-		trailingIdx++
-		for i := trailingIdx; i <= today; i++ {
-			tmp := inLow[i]
-			if tmp < lowest {
-				lowest = tmp
-			}
-			tmp = inHigh[i]
-			if tmp > highest {
-				highest = tmp
-			}
-		}
-		outReal[outIdx] = (highest + lowest) / 2.0
-		outIdx++
-		today++
-	}
-	return outReal
-}
-
-// Sar - Parabolic SAR
-// real = Sar(high, low, acceleration=0, maximum=0)
-func Sar(inHigh []float64, inLow []float64, inAcceleration float64, inMaximum float64) []float64 {
-
-	outReal := make([]float64, len(inHigh))
-
-	af := inAcceleration
-	if af > inMaximum {
-		af, inAcceleration = inMaximum, inMaximum
-	}
-
-	epTemp := MinusDM(inHigh, inLow, 1)
-	isLong := 1
-	if epTemp[1] > 0 {
-		isLong = 0
-	}
-	startIdx := 1
-	outIdx := startIdx
-	todayIdx := startIdx
-	newHigh := inHigh[todayIdx-1]
-	newLow := inLow[todayIdx-1]
-	sar, ep := 0.0, 0.0
-	if isLong == 1 {
-		ep = inHigh[todayIdx]
-		sar = newLow
-	} else {
-		ep = inLow[todayIdx]
-		sar = newHigh
-	}
-	newLow = inLow[todayIdx]
-	newHigh = inHigh[todayIdx]
-	prevLow := 0.0
-	prevHigh := 0.0
-	for todayIdx < len(inHigh) {
-		prevLow = newLow
-		prevHigh = newHigh
-		newLow = inLow[todayIdx]
-		newHigh = inHigh[todayIdx]
-		todayIdx++
-		if isLong == 1 {
-			if newLow <= sar {
-				isLong = 0
-				sar = ep
-				if sar < prevHigh {
-					sar = prevHigh
-				}
-				if sar < newHigh {
-					sar = newHigh
-				}
-				outReal[outIdx] = sar
-				outIdx++
-				af = inAcceleration
-				ep = newLow
-				sar = sar + af*(ep-sar)
-				if sar < prevHigh {
-					sar = prevHigh
-				}
-				if sar < newHigh {
-					sar = newHigh
-				}
-			} else {
-				outReal[outIdx] = sar
-				outIdx++
-				if newHigh > ep {
-					ep = newHigh
-					af += inAcceleration
-					if af > inMaximum {
-						af = inMaximum
-					}
-				}
-				sar = sar + af*(ep-sar)
-				if sar > prevLow {
-					sar = prevLow
-				}
-				if sar > newLow {
-					sar = newLow
-				}
-			}
-		} else {
-			if newHigh >= sar {
-				isLong = 1
-				sar = ep
-				if sar > prevLow {
-					sar = prevLow
-				}
-				if sar > newLow {
-					sar = newLow
-				}
-				outReal[outIdx] = sar
-				outIdx++
-				af = inAcceleration
-				ep = newHigh
-				sar = sar + af*(ep-sar)
-				if sar > prevLow {
-					sar = prevLow
-				}
-				if sar > newLow {
-					sar = newLow
-				}
-			} else {
-				outReal[outIdx] = sar
-				outIdx++
-				if newLow < ep {
-					ep = newLow
-					af += inAcceleration
-					if af > inMaximum {
-						af = inMaximum
-					}
-				}
-				sar = sar + af*(ep-sar)
-				if sar < prevHigh {
-					sar = prevHigh
-				}
-				if sar < newHigh {
-					sar = newHigh
-				}
-			}
-		}
-	}
-	return outReal
-}
-
-// SarExt - Parabolic SAR - Extended
-// real = SAREXT(high, low, startvalue=0, offsetonreverse=0, accelerationinitlong=0, accelerationlong=0, accelerationmaxlong=0, accelerationinitshort=0, accelerationshort=0, accelerationmaxshort=0)
-func SarExt(inHigh []float64, inLow []float64,
-	inStartValue float64,
-	inOffsetOnReverse float64,
-	inAccelerationInitLong float64,
-	inAccelerationLong float64,
-	inAccelerationMaxLong float64,
-	inAccelerationInitShort float64,
-	inAccelerationShort float64,
-	inAccelerationMaxShort float64) []float64 {
-
-	outReal := make([]float64, len(inHigh))
-
-	startIdx := 1
-	afLong := inAccelerationInitLong
-	afShort := inAccelerationInitShort
-	if afLong > inAccelerationMaxLong {
-		afLong, inAccelerationInitLong = inAccelerationMaxLong, inAccelerationMaxLong
-	}
-
-	if inAccelerationLong > inAccelerationMaxLong {
-		inAccelerationLong = inAccelerationMaxLong
-	}
-
-	if afShort > inAccelerationMaxShort {
-		afShort, inAccelerationInitShort = inAccelerationMaxShort, inAccelerationMaxShort
-	}
-
-	if inAccelerationShort > inAccelerationMaxShort {
-		inAccelerationShort = inAccelerationMaxShort
-	}
-
-	isLong := 0
-	if inStartValue == 0 {
-		epTemp := MinusDM(inHigh, inLow, 1)
-		if epTemp[1] > 0 {
-			isLong = 0
-		} else {
-			isLong = 1
-		}
-	} else if inStartValue > 0 {
-		isLong = 1
-	}
-	outIdx := startIdx
-	todayIdx := startIdx
-	newHigh := inHigh[todayIdx-1]
-	newLow := inLow[todayIdx-1]
-	ep := 0.0
-	sar := 0.0
-	if inStartValue == 0 {
-		if isLong == 1 {
-			ep = inHigh[todayIdx]
-			sar = newLow
-		} else {
-			ep = inLow[todayIdx]
-			sar = newHigh
-		}
-	} else if inStartValue > 0 {
-		ep = inHigh[todayIdx]
-		sar = inStartValue
-	} else {
-		ep = inLow[todayIdx]
-		sar = math.Abs(inStartValue)
-	}
-	newLow = inLow[todayIdx]
-	newHigh = inHigh[todayIdx]
-	prevLow := 0.0
-	prevHigh := 0.0
-	for todayIdx < len(inHigh) {
-		prevLow = newLow
-		prevHigh = newHigh
-		newLow = inLow[todayIdx]
-		newHigh = inHigh[todayIdx]
-		todayIdx++
-		if isLong == 1 {
-			if newLow <= sar {
-				isLong = 0
-				sar = ep
-				if sar < prevHigh {
-					sar = prevHigh
-				}
-				if sar < newHigh {
-					sar = newHigh
-				}
-				if inOffsetOnReverse != 0.0 {
-					sar += sar * inOffsetOnReverse
-				}
-				outReal[outIdx] = -sar
-				outIdx++
-				afShort = inAccelerationInitShort
-				ep = newLow
-				sar = sar + afShort*(ep-sar)
-				if sar < prevHigh {
-					sar = prevHigh
-				}
-				if sar < newHigh {
-					sar = newHigh
-				}
-			} else {
-				outReal[outIdx] = sar
-				outIdx++
-				if newHigh > ep {
-					ep = newHigh
-					afLong += inAccelerationLong
-					if afLong > inAccelerationMaxLong {
-						afLong = inAccelerationMaxLong
-					}
-				}
-				sar = sar + afLong*(ep-sar)
-				if sar > prevLow {
-					sar = prevLow
-				}
-				if sar > newLow {
-					sar = newLow
-				}
-			}
-		} else {
-			if newHigh >= sar {
-				isLong = 1
-				sar = ep
-				if sar > prevLow {
-					sar = prevLow
-				}
-				if sar > newLow {
-					sar = newLow
-				}
-				if inOffsetOnReverse != 0.0 {
-					sar -= sar * inOffsetOnReverse
-				}
-				outReal[outIdx] = sar
-				outIdx++
-				afLong = inAccelerationInitLong
-				ep = newHigh
-				sar = sar + afLong*(ep-sar)
-				if sar > prevLow {
-					sar = prevLow
-				}
-				if sar > newLow {
-					sar = newLow
-				}
-			} else {
-				outReal[outIdx] = -sar
-				outIdx++
-				if newLow < ep {
-					ep = newLow
-					afShort += inAccelerationShort
-					if afShort > inAccelerationMaxShort {
-						afShort = inAccelerationMaxShort
-					}
-				}
-				sar = sar + afShort*(ep-sar)
-				if sar < prevHigh {
-					sar = prevHigh
-				}
-				if sar < newHigh {
-					sar = newHigh
-				}
-			}
-		}
-	}
-	return outReal
-}
-
-// Sma - Simple Moving Average
-func Sma(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	lookbackTotal := inTimePeriod - 1
-	startIdx := lookbackTotal
-	periodTotal := 0.0
-	trailingIdx := startIdx - lookbackTotal
-	i := trailingIdx
-	if inTimePeriod > 1 {
-		for i < startIdx {
-			periodTotal += inReal[i]
-			i++
-		}
-	}
-	outIdx := startIdx
-	for ok := true; ok; {
-		periodTotal += inReal[i]
-		tempReal := periodTotal
-		periodTotal -= inReal[trailingIdx]
-		outReal[outIdx] = tempReal / float64(inTimePeriod)
-		trailingIdx++
-		i++
-		outIdx++
-		ok = i < len(outReal)
-	}
-
-	return outReal
-}
-
-// T3 - Triple Exponential Moving Average (T3) (lookback=6*inTimePeriod)
-func T3(inReal []float64, inTimePeriod int, inVFactor float64) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	lookbackTotal := 6 * (inTimePeriod - 1)
-	startIdx := lookbackTotal
-	today := startIdx - lookbackTotal
-	k := 2.0 / (float64(inTimePeriod) + 1.0)
-	oneMinusK := 1.0 - k
-	tempReal := inReal[today]
-	today++
-	for i := inTimePeriod - 1; i > 0; i-- {
-		tempReal += inReal[today]
-		today++
-	}
-	e1 := tempReal / float64(inTimePeriod)
-	tempReal = e1
-	for i := inTimePeriod - 1; i > 0; i-- {
-		e1 = (k * inReal[today]) + (oneMinusK * e1)
-		tempReal += e1
-		today++
-	}
-	e2 := tempReal / float64(inTimePeriod)
-	tempReal = e2
-	for i := inTimePeriod - 1; i > 0; i-- {
-		e1 = (k * inReal[today]) + (oneMinusK * e1)
-		e2 = (k * e1) + (oneMinusK * e2)
-		tempReal += e2
-		today++
-	}
-	e3 := tempReal / float64(inTimePeriod)
-	tempReal = e3
-	for i := inTimePeriod - 1; i > 0; i-- {
-		e1 = (k * inReal[today]) + (oneMinusK * e1)
-		e2 = (k * e1) + (oneMinusK * e2)
-		e3 = (k * e2) + (oneMinusK * e3)
-		tempReal += e3
-		today++
-	}
-	e4 := tempReal / float64(inTimePeriod)
-	tempReal = e4
-	for i := inTimePeriod - 1; i > 0; i-- {
-		e1 = (k * inReal[today]) + (oneMinusK * e1)
-		e2 = (k * e1) + (oneMinusK * e2)
-		e3 = (k * e2) + (oneMinusK * e3)
-		e4 = (k * e3) + (oneMinusK * e4)
-		tempReal += e4
-		today++
-	}
-	e5 := tempReal / float64(inTimePeriod)
-	tempReal = e5
-	for i := inTimePeriod - 1; i > 0; i-- {
-		e1 = (k * inReal[today]) + (oneMinusK * e1)
-		e2 = (k * e1) + (oneMinusK * e2)
-		e3 = (k * e2) + (oneMinusK * e3)
-		e4 = (k * e3) + (oneMinusK * e4)
-		e5 = (k * e4) + (oneMinusK * e5)
-		tempReal += e5
-		today++
-	}
-	e6 := tempReal / float64(inTimePeriod)
-	for today <= startIdx {
-		e1 = (k * inReal[today]) + (oneMinusK * e1)
-		e2 = (k * e1) + (oneMinusK * e2)
-		e3 = (k * e2) + (oneMinusK * e3)
-		e4 = (k * e3) + (oneMinusK * e4)
-		e5 = (k * e4) + (oneMinusK * e5)
-		e6 = (k * e5) + (oneMinusK * e6)
-		today++
-	}
-	tempReal = inVFactor * inVFactor
-	c1 := -(tempReal * inVFactor)
-	c2 := 3.0 * (tempReal - c1)
-	c3 := -6.0*tempReal - 3.0*(inVFactor-c1)
-	c4 := 1.0 + 3.0*inVFactor - c1 + 3.0*tempReal
-	outIdx := lookbackTotal
-	outReal[outIdx] = c1*e6 + c2*e5 + c3*e4 + c4*e3
-	outIdx++
-	for today < len(inReal) {
-		e1 = (k * inReal[today]) + (oneMinusK * e1)
-		e2 = (k * e1) + (oneMinusK * e2)
-		e3 = (k * e2) + (oneMinusK * e3)
-		e4 = (k * e3) + (oneMinusK * e4)
-		e5 = (k * e4) + (oneMinusK * e5)
-		e6 = (k * e5) + (oneMinusK * e6)
-		outReal[outIdx] = c1*e6 + c2*e5 + c3*e4 + c4*e3
-		outIdx++
-		today++
-	}
-
-	return outReal
-}
-
-// Tema - Triple Exponential Moving Average
-func Tema(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-	firstEMA := Ema(inReal, inTimePeriod)
-	secondEMA := Ema(firstEMA[inTimePeriod-1:], inTimePeriod)
-	thirdEMA := Ema(secondEMA[inTimePeriod-1:], inTimePeriod)
-
-	outIdx := (inTimePeriod * 3) - 3
-	secondEMAIdx := (inTimePeriod * 2) - 2
-	thirdEMAIdx := inTimePeriod - 1
-
-	for outIdx < len(inReal) {
-		outReal[outIdx] = thirdEMA[thirdEMAIdx] + ((3.0 * firstEMA[outIdx]) - (3.0 * secondEMA[secondEMAIdx]))
-		outIdx++
-		secondEMAIdx++
-		thirdEMAIdx++
-	}
-
-	return outReal
-}
-
-// Trima - Triangular Moving Average
-func Trima(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	lookbackTotal := inTimePeriod - 1
-	startIdx := lookbackTotal
-	outIdx := inTimePeriod - 1
-	var factor float64
-
-	if inTimePeriod%2 == 1 {
-		i := inTimePeriod >> 1
-		factor = (float64(i) + 1.0) * (float64(i) + 1.0)
-		factor = 1.0 / factor
-		trailingIdx := startIdx - lookbackTotal
-		middleIdx := trailingIdx + i
-		todayIdx := middleIdx + i
-		numerator := 0.0
-		numeratorSub := 0.0
-		for i := middleIdx; i >= trailingIdx; i-- {
-			tempReal := inReal[i]
-			numeratorSub += tempReal
-			numerator += numeratorSub
-		}
-		numeratorAdd := 0.0
-		middleIdx++
-		for i := middleIdx; i <= todayIdx; i++ {
-			tempReal := inReal[i]
-			numeratorAdd += tempReal
-			numerator += numeratorAdd
-		}
-		outIdx = inTimePeriod - 1
-		tempReal := inReal[trailingIdx]
-		trailingIdx++
-		outReal[outIdx] = numerator * factor
-		outIdx++
-		todayIdx++
-		for todayIdx < len(inReal) {
-			numerator -= numeratorSub
-			numeratorSub -= tempReal
-			tempReal = inReal[middleIdx]
-			middleIdx++
-			numeratorSub += tempReal
-			numerator += numeratorAdd
-			numeratorAdd -= tempReal
-			tempReal = inReal[todayIdx]
-			todayIdx++
-			numeratorAdd += tempReal
-			numerator += tempReal
-			tempReal = inReal[trailingIdx]
-			trailingIdx++
-			outReal[outIdx] = numerator * factor
-			outIdx++
-		}
-
-	} else {
-
-		i := (inTimePeriod >> 1)
-		factor = float64(i) * (float64(i) + 1)
-		factor = 1.0 / factor
-		trailingIdx := startIdx - lookbackTotal
-		middleIdx := trailingIdx + i - 1
-		todayIdx := middleIdx + i
-		numerator := 0.0
-		numeratorSub := 0.0
-		for i := middleIdx; i >= trailingIdx; i-- {
-			tempReal := inReal[i]
-			numeratorSub += tempReal
-			numerator += numeratorSub
-		}
-		numeratorAdd := 0.0
-		middleIdx++
-		for i := middleIdx; i <= todayIdx; i++ {
-			tempReal := inReal[i]
-			numeratorAdd += tempReal
-			numerator += numeratorAdd
-		}
-		outIdx = inTimePeriod - 1
-		tempReal := inReal[trailingIdx]
-		trailingIdx++
-		outReal[outIdx] = numerator * factor
-		outIdx++
-		todayIdx++
-
-		for todayIdx < len(inReal) {
-			numerator -= numeratorSub
-			numeratorSub -= tempReal
-			tempReal = inReal[middleIdx]
-			middleIdx++
-			numeratorSub += tempReal
-			numeratorAdd -= tempReal
-			numerator += numeratorAdd
-			tempReal = inReal[todayIdx]
-			todayIdx++
-			numeratorAdd += tempReal
-			numerator += tempReal
-			tempReal = inReal[trailingIdx]
-			trailingIdx++
-			outReal[outIdx] = numerator * factor
-			outIdx++
-		}
-	}
-	return outReal
-}
-
-// Wma - Weighted Moving Average
-func Wma(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	lookbackTotal := inTimePeriod - 1
-	startIdx := lookbackTotal
-
-	if inTimePeriod == 1 {
-		copy(outReal, inReal)
-		return outReal
-	}
-	divider := (inTimePeriod * (inTimePeriod + 1)) >> 1
-	outIdx := inTimePeriod - 1
-	trailingIdx := startIdx - lookbackTotal
-	periodSum, periodSub := 0.0, 0.0
-	inIdx := trailingIdx
-	i := 1
-	for inIdx < startIdx {
-		tempReal := inReal[inIdx]
-		periodSub += tempReal
-		periodSum += tempReal * float64(i)
-		inIdx++
-		i++
-	}
-	trailingValue := 0.0
-	for inIdx < len(inReal) {
-		tempReal := inReal[inIdx]
-		periodSub += tempReal
-		periodSub -= trailingValue
-		periodSum += tempReal * float64(inTimePeriod)
-		trailingValue = inReal[trailingIdx]
-		outReal[outIdx] = periodSum / float64(divider)
-		periodSum -= periodSub
-		inIdx++
-		trailingIdx++
-		outIdx++
-	}
-	return outReal
-}
 
 /* Momentum Indicators */
 
@@ -1975,93 +548,6 @@ func Dx(inHigh []float64, inLow []float64, inClose []float64, inTimePeriod int) 
 		outIdx++
 	}
 	return outReal
-}
-
-// Macd - Moving Average Convergence/Divergence
-// unstable period ~= 100
-func Macd(inReal []float64, inFastPeriod int, inSlowPeriod int, inSignalPeriod int) ([]float64, []float64, []float64) {
-
-	if inSlowPeriod < inFastPeriod {
-		inSlowPeriod, inFastPeriod = inFastPeriod, inSlowPeriod
-	}
-
-	k1 := 0.0
-	k2 := 0.0
-	if inSlowPeriod != 0 {
-		k1 = 2.0 / float64(inSlowPeriod+1)
-	} else {
-		inSlowPeriod = 26
-		k1 = 0.075
-	}
-	if inFastPeriod != 0 {
-		k2 = 2.0 / float64(inFastPeriod+1)
-	} else {
-		inFastPeriod = 12
-		k2 = 0.15
-	}
-
-	lookbackSignal := inSignalPeriod - 1
-	lookbackTotal := lookbackSignal
-	lookbackTotal += (inSlowPeriod - 1)
-
-	fastEMABuffer := ema(inReal, inFastPeriod, k2)
-	slowEMABuffer := ema(inReal, inSlowPeriod, k1)
-	for i := 0; i < len(fastEMABuffer); i++ {
-		fastEMABuffer[i] = fastEMABuffer[i] - slowEMABuffer[i]
-	}
-
-	outMACD := make([]float64, len(inReal))
-	for i := lookbackTotal - 1; i < len(fastEMABuffer); i++ {
-		outMACD[i] = fastEMABuffer[i]
-	}
-	outMACDSignal := ema(outMACD, inSignalPeriod, (2.0 / float64(inSignalPeriod+1)))
-
-	outMACDHist := make([]float64, len(inReal))
-	for i := lookbackTotal; i < len(outMACDHist); i++ {
-		outMACDHist[i] = outMACD[i] - outMACDSignal[i]
-	}
-
-	return outMACD, outMACDSignal, outMACDHist
-}
-
-// MacdExt - MACD with controllable MA type
-// unstable period ~= 100
-func MacdExt(inReal []float64, inFastPeriod int, inFastMAType MaType, inSlowPeriod int, inSlowMAType MaType, inSignalPeriod int, inSignalMAType MaType) ([]float64, []float64, []float64) {
-
-	lookbackLargest := 0
-	if inFastPeriod < inSlowPeriod {
-		lookbackLargest = inSlowPeriod
-	} else {
-		lookbackLargest = inFastPeriod
-	}
-	lookbackTotal := (inSignalPeriod - 1) + (lookbackLargest - 1)
-
-	outMACD := make([]float64, len(inReal))
-	outMACDSignal := make([]float64, len(inReal))
-	outMACDHist := make([]float64, len(inReal))
-
-	slowMABuffer := Ma(inReal, inSlowPeriod, inSlowMAType)
-	fastMABuffer := Ma(inReal, inFastPeriod, inFastMAType)
-	tempBuffer1 := make([]float64, len(inReal))
-
-	for i := 0; i < len(slowMABuffer); i++ {
-		tempBuffer1[i] = fastMABuffer[i] - slowMABuffer[i]
-	}
-	tempBuffer2 := Ma(tempBuffer1, inSignalPeriod, inSignalMAType)
-
-	for i := lookbackTotal; i < len(outMACDHist); i++ {
-		outMACD[i] = tempBuffer1[i]
-		outMACDSignal[i] = tempBuffer2[i]
-		outMACDHist[i] = outMACD[i] - outMACDSignal[i]
-	}
-
-	return outMACD, outMACDSignal, outMACDHist
-}
-
-// MacdFix - MACD Fix 12/26
-// unstable period ~= 100
-func MacdFix(inReal []float64, inSignalPeriod int) ([]float64, []float64, []float64) {
-	return Macd(inReal, 0, 0, inSignalPeriod)
 }
 
 // MinusDI - Minus Directional Indicator
@@ -2800,96 +1286,6 @@ func Rocr100(inReal []float64, inTimePeriod int) []float64 {
 		outIdx++
 		inIdx++
 	}
-	return outReal
-}
-
-// Rsi - Relative strength index
-func Rsi(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	if inTimePeriod < 2 {
-		return outReal
-	}
-
-	// variable declarations
-	tempValue1 := 0.0
-	tempValue2 := 0.0
-	outIdx := inTimePeriod
-	today := 0
-	prevValue := inReal[today]
-	prevGain := 0.0
-	prevLoss := 0.0
-	today++
-
-	for i := inTimePeriod; i > 0; i-- {
-		tempValue1 = inReal[today]
-		today++
-		tempValue2 = tempValue1 - prevValue
-		prevValue = tempValue1
-		if tempValue2 < 0 {
-			prevLoss -= tempValue2
-		} else {
-			prevGain += tempValue2
-		}
-	}
-
-	prevLoss /= float64(inTimePeriod)
-	prevGain /= float64(inTimePeriod)
-
-	if today > 0 {
-
-		tempValue1 = prevGain + prevLoss
-		if !((-0.00000000000001 < tempValue1) && (tempValue1 < 0.00000000000001)) {
-			outReal[outIdx] = 100.0 * (prevGain / tempValue1)
-		} else {
-			outReal[outIdx] = 0.0
-		}
-		outIdx++
-
-	} else {
-
-		for today < 0 {
-			tempValue1 = inReal[today]
-			tempValue2 = tempValue1 - prevValue
-			prevValue = tempValue1
-			prevLoss *= float64(inTimePeriod - 1)
-			prevGain *= float64(inTimePeriod - 1)
-			if tempValue2 < 0 {
-				prevLoss -= tempValue2
-			} else {
-				prevGain += tempValue2
-			}
-			prevLoss /= float64(inTimePeriod)
-			prevGain /= float64(inTimePeriod)
-			today++
-		}
-	}
-
-	for today < len(inReal) {
-
-		tempValue1 = inReal[today]
-		today++
-		tempValue2 = tempValue1 - prevValue
-		prevValue = tempValue1
-		prevLoss *= float64(inTimePeriod - 1)
-		prevGain *= float64(inTimePeriod - 1)
-		if tempValue2 < 0 {
-			prevLoss -= tempValue2
-		} else {
-			prevGain += tempValue2
-		}
-		prevLoss /= float64(inTimePeriod)
-		prevGain /= float64(inTimePeriod)
-		tempValue1 = prevGain + prevLoss
-		if !((-0.00000000000001 < tempValue1) && (tempValue1 < 0.00000000000001)) {
-			outReal[outIdx] = 100.0 * (prevGain / tempValue1)
-		} else {
-			outReal[outIdx] = 0.0
-		}
-		outIdx++
-	}
-
 	return outReal
 }
 
@@ -5360,507 +3756,6 @@ func Var(inReal []float64, inTimePeriod int) []float64 {
 		outIdx++
 		ok = i < len(inReal)
 	}
-	return outReal
-}
-
-/* Math Transform Functions */
-
-// Acos - Vector Trigonometric ACOS
-func Acos(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Acos(inReal[i])
-	}
-	return outReal
-}
-
-// Asin - Vector Trigonometric ASIN
-func Asin(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Asin(inReal[i])
-	}
-	return outReal
-}
-
-// Atan - Vector Trigonometric ATAN
-func Atan(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Atan(inReal[i])
-	}
-	return outReal
-}
-
-// Ceil - Vector CEIL
-func Ceil(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Ceil(inReal[i])
-	}
-	return outReal
-}
-
-// Cos - Vector Trigonometric COS
-func Cos(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Cos(inReal[i])
-	}
-	return outReal
-}
-
-// Cosh - Vector Trigonometric COSH
-func Cosh(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Cosh(inReal[i])
-	}
-	return outReal
-}
-
-// Exp - Vector atrithmetic EXP
-func Exp(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Exp(inReal[i])
-	}
-	return outReal
-}
-
-// Floor - Vector FLOOR
-func Floor(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Floor(inReal[i])
-	}
-	return outReal
-}
-
-// Ln - Vector natural log LN
-func Ln(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Log(inReal[i])
-	}
-	return outReal
-}
-
-// Log10 - Vector LOG10
-func Log10(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Log10(inReal[i])
-	}
-	return outReal
-}
-
-// Sin - Vector Trigonometric SIN
-func Sin(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Sin(inReal[i])
-	}
-	return outReal
-}
-
-// Sinh - Vector Trigonometric SINH
-func Sinh(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Sinh(inReal[i])
-	}
-	return outReal
-}
-
-// Sqrt - Vector SQRT
-func Sqrt(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Sqrt(inReal[i])
-	}
-	return outReal
-}
-
-// Tan - Vector Trigonometric TAN
-func Tan(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Tan(inReal[i])
-	}
-	return outReal
-}
-
-// Tanh - Vector Trigonometric TANH
-func Tanh(inReal []float64) []float64 {
-	outReal := make([]float64, len(inReal))
-	for i := 0; i < len(inReal); i++ {
-		outReal[i] = math.Tanh(inReal[i])
-	}
-	return outReal
-}
-
-/* Math Operator Functions */
-
-// Add - Vector arithmetic addition
-func Add(inReal0 []float64, inReal1 []float64) []float64 {
-	outReal := make([]float64, len(inReal0))
-	for i := 0; i < len(inReal0); i++ {
-		outReal[i] = inReal0[i] + inReal1[i]
-	}
-	return outReal
-}
-
-// Div - Vector arithmetic division
-func Div(inReal0 []float64, inReal1 []float64) []float64 {
-	outReal := make([]float64, len(inReal0))
-	for i := 0; i < len(inReal0); i++ {
-		outReal[i] = inReal0[i] / inReal1[i]
-	}
-	return outReal
-}
-
-// Max - Highest value over a period
-func Max(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	if inTimePeriod < 2 {
-		return outReal
-	}
-
-	nbInitialElementNeeded := inTimePeriod - 1
-	startIdx := nbInitialElementNeeded
-	outIdx := startIdx
-	today := startIdx
-	trailingIdx := startIdx - nbInitialElementNeeded
-	highestIdx := -1
-	highest := 0.0
-
-	for today < len(outReal) {
-
-		tmp := inReal[today]
-
-		if highestIdx < trailingIdx {
-			highestIdx = trailingIdx
-			highest = inReal[highestIdx]
-			i := highestIdx + 1
-			for i <= today {
-				tmp = inReal[i]
-				if tmp > highest {
-					highestIdx = i
-					highest = tmp
-				}
-				i++
-			}
-		} else if tmp >= highest {
-			highestIdx = today
-			highest = tmp
-		}
-		outReal[outIdx] = highest
-		outIdx++
-		trailingIdx++
-		today++
-	}
-
-	return outReal
-}
-
-// MaxIndex - Index of highest value over a specified period
-func MaxIndex(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	if inTimePeriod < 2 {
-		return outReal
-	}
-
-	nbInitialElementNeeded := inTimePeriod - 1
-	startIdx := nbInitialElementNeeded
-	outIdx := startIdx
-	today := startIdx
-	trailingIdx := startIdx - nbInitialElementNeeded
-	highestIdx := -1
-	highest := 0.0
-	for today < len(inReal) {
-		tmp := inReal[today]
-		if highestIdx < trailingIdx {
-			highestIdx = trailingIdx
-			highest = inReal[highestIdx]
-			i := highestIdx + 1
-			for i <= today {
-				tmp := inReal[i]
-				if tmp > highest {
-					highestIdx = i
-					highest = tmp
-				}
-				i++
-			}
-		} else if tmp >= highest {
-			highestIdx = today
-			highest = tmp
-		}
-		outReal[outIdx] = float64(highestIdx)
-		outIdx++
-		trailingIdx++
-		today++
-	}
-
-	return outReal
-}
-
-// Min - Lowest value over a period
-func Min(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	if inTimePeriod < 2 {
-		return outReal
-	}
-
-	nbInitialElementNeeded := inTimePeriod - 1
-	startIdx := nbInitialElementNeeded
-	outIdx := startIdx
-	today := startIdx
-	trailingIdx := startIdx - nbInitialElementNeeded
-	lowestIdx := -1
-	lowest := 0.0
-	for today < len(outReal) {
-
-		tmp := inReal[today]
-
-		if lowestIdx < trailingIdx {
-			lowestIdx = trailingIdx
-			lowest = inReal[lowestIdx]
-			i := lowestIdx + 1
-			for i <= today {
-				tmp = inReal[i]
-				if tmp < lowest {
-					lowestIdx = i
-					lowest = tmp
-				}
-				i++
-			}
-		} else if tmp <= lowest {
-			lowestIdx = today
-			lowest = tmp
-		}
-		outReal[outIdx] = lowest
-		outIdx++
-		trailingIdx++
-		today++
-	}
-
-	return outReal
-}
-
-// MinIndex - Index of lowest value over a specified period
-func MinIndex(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	if inTimePeriod < 2 {
-		return outReal
-	}
-
-	nbInitialElementNeeded := inTimePeriod - 1
-	startIdx := nbInitialElementNeeded
-	outIdx := startIdx
-	today := startIdx
-	trailingIdx := startIdx - nbInitialElementNeeded
-	lowestIdx := -1
-	lowest := 0.0
-	for today < len(inReal) {
-		tmp := inReal[today]
-		if lowestIdx < trailingIdx {
-			lowestIdx = trailingIdx
-			lowest = inReal[lowestIdx]
-			i := lowestIdx + 1
-			for i <= today {
-				tmp = inReal[i]
-				if tmp < lowest {
-					lowestIdx = i
-					lowest = tmp
-				}
-				i++
-			}
-		} else if tmp <= lowest {
-			lowestIdx = today
-			lowest = tmp
-		}
-		outReal[outIdx] = float64(lowestIdx)
-		outIdx++
-		trailingIdx++
-		today++
-	}
-	return outReal
-}
-
-// MinMax - Lowest and highest values over a specified period
-func MinMax(inReal []float64, inTimePeriod int) ([]float64, []float64) {
-
-	outMin := make([]float64, len(inReal))
-	outMax := make([]float64, len(inReal))
-
-	nbInitialElementNeeded := (inTimePeriod - 1)
-	startIdx := nbInitialElementNeeded
-	outIdx := startIdx
-	today := startIdx
-	trailingIdx := startIdx - nbInitialElementNeeded
-	highestIdx := -1
-	highest := 0.0
-	lowestIdx := -1
-	lowest := 0.0
-	for today < len(inReal) {
-		tmpLow, tmpHigh := inReal[today], inReal[today]
-		if highestIdx < trailingIdx {
-			highestIdx = trailingIdx
-			highest = inReal[highestIdx]
-			i := highestIdx
-			i++
-			for i <= today {
-				tmpHigh = inReal[i]
-				if tmpHigh > highest {
-					highestIdx = i
-					highest = tmpHigh
-				}
-				i++
-			}
-		} else if tmpHigh >= highest {
-			highestIdx = today
-			highest = tmpHigh
-		}
-		if lowestIdx < trailingIdx {
-			lowestIdx = trailingIdx
-			lowest = inReal[lowestIdx]
-			i := lowestIdx
-			i++
-			for i <= today {
-				tmpLow = inReal[i]
-				if tmpLow < lowest {
-					lowestIdx = i
-					lowest = tmpLow
-				}
-				i++
-			}
-		} else if tmpLow <= lowest {
-			lowestIdx = today
-			lowest = tmpLow
-		}
-		outMax[outIdx] = highest
-		outMin[outIdx] = lowest
-		outIdx++
-		trailingIdx++
-		today++
-	}
-	return outMin, outMax
-}
-
-// MinMaxIndex - Indexes of lowest and highest values over a specified period
-func MinMaxIndex(inReal []float64, inTimePeriod int) ([]float64, []float64) {
-
-	outMinIdx := make([]float64, len(inReal))
-	outMaxIdx := make([]float64, len(inReal))
-
-	nbInitialElementNeeded := (inTimePeriod - 1)
-	startIdx := nbInitialElementNeeded
-	outIdx := startIdx
-	today := startIdx
-	trailingIdx := startIdx - nbInitialElementNeeded
-	highestIdx := -1
-	highest := 0.0
-	lowestIdx := -1
-	lowest := 0.0
-	for today < len(inReal) {
-		tmpLow, tmpHigh := inReal[today], inReal[today]
-		if highestIdx < trailingIdx {
-			highestIdx = trailingIdx
-			highest = inReal[highestIdx]
-			i := highestIdx
-			i++
-			for i <= today {
-				tmpHigh = inReal[i]
-				if tmpHigh > highest {
-					highestIdx = i
-					highest = tmpHigh
-				}
-				i++
-			}
-		} else if tmpHigh >= highest {
-			highestIdx = today
-			highest = tmpHigh
-		}
-		if lowestIdx < trailingIdx {
-			lowestIdx = trailingIdx
-			lowest = inReal[lowestIdx]
-			i := lowestIdx
-			i++
-			for i <= today {
-				tmpLow = inReal[i]
-				if tmpLow < lowest {
-					lowestIdx = i
-					lowest = tmpLow
-				}
-				i++
-			}
-		} else if tmpLow <= lowest {
-			lowestIdx = today
-			lowest = tmpLow
-		}
-		outMaxIdx[outIdx] = float64(highestIdx)
-		outMinIdx[outIdx] = float64(lowestIdx)
-		outIdx++
-		trailingIdx++
-		today++
-	}
-	return outMinIdx, outMaxIdx
-}
-
-// Mult - Vector arithmetic multiply
-func Mult(inReal0 []float64, inReal1 []float64) []float64 {
-	outReal := make([]float64, len(inReal0))
-	for i := 0; i < len(inReal0); i++ {
-		outReal[i] = inReal0[i] * inReal1[i]
-	}
-	return outReal
-}
-
-// Sub - Vector arithmetic subtraction
-func Sub(inReal0 []float64, inReal1 []float64) []float64 {
-	outReal := make([]float64, len(inReal0))
-	for i := 0; i < len(inReal0); i++ {
-		outReal[i] = inReal0[i] - inReal1[i]
-	}
-	return outReal
-}
-
-// Sum - Vector summation
-func Sum(inReal []float64, inTimePeriod int) []float64 {
-
-	outReal := make([]float64, len(inReal))
-
-	lookbackTotal := inTimePeriod - 1
-	startIdx := lookbackTotal
-	periodTotal := 0.0
-	trailingIdx := startIdx - lookbackTotal
-	i := trailingIdx
-	if inTimePeriod > 1 {
-		for i < startIdx {
-			periodTotal += inReal[i]
-			i++
-		}
-	}
-	outIdx := startIdx
-	for i < len(inReal) {
-		periodTotal += inReal[i]
-		tempReal := periodTotal
-		periodTotal -= inReal[trailingIdx]
-		outReal[outIdx] = tempReal
-		i++
-		trailingIdx++
-		outIdx++
-	}
-
 	return outReal
 }
