+++ indicators/talib_test.go
@@ -4,7 +4,7 @@ Copyright 2018 Alessandro Sanino
 Licensed under terms of MIT license (see LICENSE)
 */
 
-package talib
+package indicators
 
 import (
 	"fmt"
@@ -18,9 +18,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/stretchr/testify/require"
-
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func ok(t *testing.T, err error) {
